+++ slack-service/src/main/java/com/business/slackservice/application/dto/request/slackEventType/ReqSlackEventTypePostDTOApiV1.java
@@ -1,6 +1,6 @@
 package com.business.slackservice.application.dto.request.slackEventType;
 
-import com.business.slackservice.domain.slack.entity.SlackEventTypeEntity;
+import com.business.slackservice.domain.slackEventType.entity.SlackEventTypeEntity;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
