--- slack-service/src/main/java/com/business/slackservice/domain/slack/repository/SlackTemplateRepository.java
@@ -1,5 +1,0 @@
-package com.business.slackservice.domain.slack.repository;
-
-public interface SlackTemplateRepository {
-
-}
