+++ slack-service/src/main/java/com/business/slackservice/domain/slackTemplate/repository/SlackTemplateRepository.java
@@ -1,0 +1,9 @@
+package com.business.slackservice.domain.slackTemplate.repository;
+
+import com.business.slackservice.domain.slackTemplate.entity.SlackTemplateEntity;
+
+public interface SlackTemplateRepository {
+
+    SlackTemplateEntity save(SlackTemplateEntity slackTemplate);
+}
+
