+++ slack-service/src/main/java/com/business/slackservice/infrastructure/persistence/SlackEventTypeJpaRepository.java
@@ -1,7 +1,7 @@
 package com.business.slackservice.infrastructure.persistence;
 
-import com.business.slackservice.domain.slack.entity.SlackEventTypeEntity;
-import com.business.slackservice.domain.slack.repository.SlackEventTypeRepository;
+import com.business.slackservice.domain.slackEventType.entity.SlackEventTypeEntity;
+import com.business.slackservice.domain.slackEventType.repository.SlackEventTypeRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
