+++ slack-service/src/main/java/com/business/slackservice/infrastructure/persistence/SlackTemplateJpaRepository.java
@@ -1,11 +1,11 @@
 package com.business.slackservice.infrastructure.persistence;
 
-import com.business.slackservice.domain.slack.entity.SlackEntity;
-import com.business.slackservice.domain.slack.repository.SlackTemplateRepository;
+import com.business.slackservice.domain.slackTemplate.entity.SlackTemplateEntity;
+import com.business.slackservice.domain.slackTemplate.repository.SlackTemplateRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SlackTemplateJpaRepository extends JpaRepository<SlackEntity, UUID>,
+public interface SlackTemplateJpaRepository extends JpaRepository<SlackTemplateEntity, UUID>,
     SlackTemplateRepository {
 
 }
